RuoYi-Vue3-FastAPI-master/ruoyi-fastapi-backend/requirements.txt
# ============================================
# RuoYi-Vue3-FastAPI 项目依赖文件
# 数据库: MySQL
# Python版本: >= 3.10
# ============================================

# ============================================
# 核心框架（必需）
# ============================================
fastapi==0.116.1
uvicorn==0.37.0
starlette==0.47.3

# ============================================
# 数据库相关（必需）
# ============================================
SQLAlchemy==2.0.43
alembic==1.16.4
asyncmy==0.2.10
PyMySQL==1.1.1

# ============================================
# Redis缓存（必需）
# ============================================
redis==6.4.0

# ============================================
# 认证和安全（必需）
# ============================================
PyJWT==2.10.1
bcrypt==5.0.0
cryptography==46.0.3

# ============================================
# 数据验证和处理（必需）
# ============================================
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
pydantic_validation_decorator==0.1.4

# ============================================
# 数据处理（Excel导出功能需要）
# ============================================
pandas==2.3.2
numpy==2.3.4
openpyxl==3.1.5

# ============================================
# 图像处理（验证码功能需要）
# ============================================
pillow==11.3.0

# ============================================
# 任务调度（定时任务功能需要）
# ============================================
APScheduler==3.11.0

# ============================================
# 日志（必需）
# ============================================
loguru==0.7.3

# ============================================
# 系统监控（必需）
# ============================================
psutil==7.0.0

# ============================================
# 用户代理解析（必需）
# ============================================
user-agents==2.2.0

# ============================================
# SQL解析（必需）
# ============================================
sqlglot==27.8.0

# ============================================
# 日期时间处理（必需）
# ============================================
python-dateutil==2.9.0.post0

# ============================================
# 异步支持（必需）
# ============================================
async-lru==2.0.5
greenlet==3.2.4

# ============================================
# HTTP客户端（必需）
# ============================================
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0

# ============================================
# 模板引擎（代码生成和数据库迁移需要）
# ============================================
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3

# ============================================
# 配置和环境（必需）
# ============================================
python-dotenv==1.1.1

# ============================================
# 类型支持（必需）
# ============================================
typing_extensions==4.15.0
annotated-types==0.7.0

# ============================================
# 文件上传支持（FastAPI文件上传需要）
# ============================================
python-multipart==0.0.20

# ============================================
# 开发工具（开发模式热重载需要）
# ============================================
watchfiles==1.1.1

# ============================================
# 其他必需依赖
# ============================================
certifi==2025.10.5
idna==3.11
sniffio==1.3.1
urllib3==2.5.0
uvloop==0.21.0
